+++ Final_Project/Final_Project/TcpServer.h
@@ -7,7 +7,7 @@ class TcpServer :public TcpSocket
 {
 public:
 	//open tcp socket on port number 'port'
-	TcpServer(int port);
+	TcpServer(unsigned short port);
 	//bind + listen + accept. throws an exception in case of failure.
 	void serve(Message* mess);// , mutex& m);
 	bool isValid()const;
